El objeto Root del layout se registra y ante un resize de la ventana, se renderiza el y sus descendientes.

+++ src/js/layout/root.js
@@ -13,6 +13,19 @@ define(['jquery', 'layout/panel', 'jqueryui'], function ($, P, _) {
 
       this._$anchor_element = $(dom_parent_element);
       this._child = null;
+
+      var self = this;
+      var last_resize_render = null;
+      $(window).resize(function () {
+         if (last_resize_render) {
+            clearTimeout(last_resize_render);
+         }
+
+         last_resize_render = setTimeout(function () {
+            self.render();
+            last_resize_render = null;
+         }, 100);
+      });
    };
 
    Root.prototype.__proto__ = Parent.prototype;
@@ -57,12 +70,16 @@ define(['jquery', 'layout/panel', 'jqueryui'], function ($, P, _) {
 
    Root.prototype.render = function () {
       //XXX ignore $box
-      this._child.box = this._$anchor_element;
-      this._child.render(this._$anchor_element);
+      if(this._child) {
+         this._child.box = this._$anchor_element;
+         this._child.render(this._$anchor_element);
+      }
    };
 
    Root.prototype.unlink = function () {
-      this._child.unlink();
+      if(this._child) {
+         this._child.unlink();
+      }
    };
 
    // Implement the method of Panel. See panel.js
